refactor: MyPage 프레젠테이셔널화, 컨테이너화
