+++ src/pages/User/MyPage.jsx
@@ -1,86 +1,22 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Avatar, Tabs, Button } from 'flowbite-react';
-import { useDispatch, useSelector } from 'react-redux';
-import { useNavigate } from 'react-router-dom';
 import UserSellList from '../../features/User/mypage/components/UserSellList';
 import UserBuyList from '../../features/User/mypage/components/UserBuyList';
 import LikeList from '../../features/User/mypage/components/LikeList';
-import { deleteUser, setUserField } from '../../features/User/store/userSlice';
-import axios from 'axios';
 import { ImageDropZone } from '../../shared/ImageDropZone';
 
-const MyPage = () => {
-  const [images, setImages] = useState([]); // 이미지 상태 관리
-  const [profileEdit, setProfileEdit] = useState(false); // 프로필 수정 상태 관리
-  const [originalProfileImage, setOriginalProfileImage] = useState(''); // 원본 프로필 이미지 저장
-  const currentUser = useSelector((state) => state.user.currentUser);
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-
-  // 컴포넌트가 마운트될 때 원래 프로필 이미지를 저장
-  useEffect(() => {
-    setOriginalProfileImage(currentUser.profile_image);
-  }, [currentUser.profile_image]);
-
-  const handleEdit = () => {
-    navigate('/mypageEdit');
-  };
-
-  // 회원 탈퇴
-  const handleDelete = async () => {
-    const isDelete = window.confirm('회원을 탈퇴하시겠습니까?');
-    if (isDelete) {
-      try {
-        const userId = currentUser.userId;
-        await axios.delete(`http://localhost:8000/user/${userId}`);
-
-        dispatch(deleteUser());
-        alert('그동안 이용해주셔서 감사합니다.');
-        navigate('/');
-      } catch (error) {
-        console.error('회원 탈퇴 중 오류가 발생했습니다:', error);
-        alert('회원 탈퇴 중 오류가 발생했습니다. 다시 시도해주세요.');
-      }
-    }
-  };
-
-  // 프로필 수정하기
-  const handleProfileClick = async () => {
-    if (profileEdit) {
-      // 수정 완료 버튼 클릭 시
-      if (images.length > 0) {
-        try {
-          const userId = currentUser.userId;
-          const newProfileImage = images[0];
-          await axios.patch(`http://localhost:8000/user/${userId}`, {
-            profile_image: newProfileImage,
-          });
-
-          // 프로필 이미지 업데이트
-          dispatch(
-            setUserField({ field: 'profile_image', value: newProfileImage }),
-          );
-          setProfileEdit(false);
-        } catch (error) {
-          console.error('프로필 수정 중 에러 발생', error);
-          alert('프로필 수정에 실패하였습니다.');
-        }
-      } else {
-        setProfileEdit(false);
-      }
-    } else {
-      // 프로필 수정하기 버튼 클릭 시
-      setProfileEdit(true);
-      setImages([]);
-    }
-  };
-
-  // 프로필 수정 시 취소 버튼 클릭
-  const handleCancel = () => {
-    setProfileEdit(false);
-    setImages([originalProfileImage]);
-  };
-
+const MyPage = ({
+  profileEdit,
+  images,
+  setImages,
+  currentUser,
+  handleProfileClick,
+  handleCancel,
+  handleEdit,
+  handleDelete,
+  sells,
+  buys,
+}) => {
   return (
     <>
       <section className="flex justify-center py-10 w-full">
@@ -127,28 +63,20 @@ const MyPage = () => {
             </Tabs.Item>
             <Tabs.Item title="판매내역">
               <div className="flex flex-wrap">
-                <div className="w-full sm:w-1/2 p-2">
-                  <UserSellList />
-                </div>
-                <div className="w-full sm:w-1/2 p-2">
-                  <UserSellList />
-                </div>
-                <div className="w-full sm:w-1/2 p-2">
-                  <UserSellList />
-                </div>
+                {sells.map((sell) => (
+                  <div className="w-full sm:w-1/2 p-2" key={sell.productId}>
+                    <UserSellList sell={sell} />
+                  </div>
+                ))}
               </div>
             </Tabs.Item>
             <Tabs.Item title="구매내역">
               <div className="flex flex-wrap">
-                <div className="w-full sm:w-1/2 p-2">
-                  <UserBuyList />
-                </div>
-                <div className="w-full sm:w-1/2 p-2">
-                  <UserBuyList />
-                </div>
-                <div className="w-full sm:w-1/2 p-2">
-                  <UserBuyList />
-                </div>
+                {buys.map((buy) => (
+                  <div className="w-full sm:w-1/2 p-2" key={buy.productId}>
+                    <UserBuyList buy={buy} />
+                  </div>
+                ))}
               </div>
             </Tabs.Item>
           </Tabs>
